+++ Magic/CHANGELOG.md
@@ -16,6 +16,7 @@
    - `to_color`: An alias for `color2` to align with EffectLib parameters
    - `particle_arrival_time`: For playing individual vibration particles
  - New external example added, `/mconfig fetch httyd` - a "How to Train Your Dragon" config pack from Night!
+ - Add `equip_on_respawn: false` option for mobs, to their equipment won't get reset on respawn/reload
 
 # 9.1.1